Optimize loading LHN dependencies
