+++ src/ggrc/assets/javascripts/controllers/lhn_controllers.js
@@ -3,7 +3,9 @@
     Licensed under http://www.apache.org/licenses/LICENSE-2.0 <see LICENSE file>
 */
 
-(function(can, $) {
+import '../models/recently_viewed_object';
+import './infinite-scroll-controller';
+
 can.Control("CMS.Controllers.LHN", {
     defaults: {
     }
@@ -385,67 +387,6 @@ can.Control("CMS.Controllers.LHN", {
     this.element.find(".bar-v").addClass("disabled");
     this.options.display_prefs.setLHNState("is_pinned", true);
   }
-});
-
-can.Control("CMS.Controllers.InfiniteScroll", {
-    defaults: {
-    }
-}, {
-    init: function() {
-    }
-
-  , " DOMMouseScroll": "prevent_overscroll"
-  , " mousewheel": "prevent_overscroll"
-  , " scroll": "prevent_overscroll"
-
-  , prevent_overscroll: function($el, ev) {
-      // Based on Troy Alford's response on StackOverflow:
-      //   http://stackoverflow.com/a/16324762
-      var scrollTop = $el[0].scrollTop
-        , scrollHeight = $el[0].scrollHeight
-        , height = $el.height()
-        , scrollTopMax = scrollHeight - height
-        , delta
-        , up
-        , loadTriggerOffset = 50
-        ;
-
-      if (ev.type === "DOMMouseScroll")
-        delta = ev.originalEvent.detail * -40;
-      else
-        delta = ev.originalEvent.wheelDelta;
-
-      up = delta > 0;
-
-      var prevent = function() {
-        ev.stopPropagation();
-        ev.preventDefault();
-        ev.returnValue = false;
-        return false;
-      }
-
-      if (ev.type === "scroll" &&
-          scrollTop > scrollTopMax - loadTriggerOffset) {
-        this.show_more($el);
-        return prevent();
-      } else if (!up && scrollTop - delta > scrollTopMax) {
-        // Scrolling down, but this will take us past the bottom.
-        $el.scrollTop(scrollHeight);
-        this.show_more($el);
-        return prevent();
-      } else if (up && delta > scrollTop) {
-        // Scrolling up, but this will take us past the top.
-        $el.scrollTop(0);
-        return prevent();
-      } else if (!up && scrollTop - delta > scrollTopMax - loadTriggerOffset) {
-        // Scrolling down, close to bottom, so start loading more
-        this.show_more($el);
-      }
-    }
-
-  , show_more: function($el) {
-      this.element.trigger("scrollNext");
-    }
 });
 
 can.Control("CMS.Controllers.LHN_Search", {
@@ -1108,149 +1049,39 @@ can.Control("CMS.Controllers.LHN_Search", {
   }
 });
 
-can.Control("CMS.Controllers.LHN_Tooltips", {
+  can.Control("GGRC.Controllers.RecentlyViewed", {
     defaults : {
-        tooltip_view: GGRC.mustache_path + "/base_objects/extended_info.mustache"
-      , trigger_selector: ".show-extended"
-      , fade_in_delay: 300
-      , fade_out_delay: 300
+      list_view : GGRC.mustache_path + "/dashboard/recently_viewed_list.mustache"
+      , max_history : 10
+      , max_display : 3
     }
-}, {
-    init: function() {
-      if (!this.options.$extended) {
-        this.options.$extended = $('#extended-info');
-        if (this.options.$extended.length < 1)
-          this.options.$extended =
-            $('<div id="extended-info" class="extended-info hide" />')
-              .appendTo('body');
-      }
-      if (!this.options.$lhs)
-        this.options.$lhs = $('#lhs');
-      // Renew event listening, since we assigned $extended, $lhs
-      this.on();
-    }
+  }, {
+    init : function() {
+      var page_model = GGRC.page_instance();
+      var instance_list = [];
+      var that = this;
 
-  // Tooltip / popover handling
-  , "{trigger_selector} mouseenter": "on_mouseenter"
-  , "{trigger_selector} mouseleave": "on_mouseleave"
-  , "{$extended} mouseleave": "on_mouseleave"
-  , "{$extended} mouseenter": "on_tooltip_mouseenter"
-
-  , on_mouseenter: function(el, ev) {
-      var instance = el.closest("[data-model]").data("model")
-                      || el.closest(":data(model)").data("model")
-        , delay = this.options.fade_in_delay
-        ;
-
-      // There isn't tooltip data available for recently viewed objects
-      if (instance instanceof GGRC.Models.RecentlyViewedObject)
-        return;
-
-      if (this.options.$extended.data('model') !== instance) {
-        clearTimeout(this.fade_in_timeout);
-        // If tooltip is already showing, show new content without delay
-        if (this.options.$extended.hasClass('in'))
-          delay = 0;
-        this.fade_in_timeout = setTimeout(
-            this.proxy('on_fade_in_timeout', el, instance), delay);
-        clearTimeout(this.fade_out_timeout);
-        this.fade_out_timeout = null;
-      } else if (this.fade_out_timeout) {
-        clearTimeout(this.fade_out_timeout);
-        this.fade_out_timeout = null;
-      }
-    }
-
-  , ensure_tooltip_visibility: function() {
-      var offset = this.options.$extended.offset().top
-        , height = this.options.$extended.height()
-        // "- 24" compensates for the Chrome URL display when hovering a link
-        // "348" should be the widht of the Chrome URL display when displaying javascript://
-        , window_height = $(window).height() + $(window).scrollTop() -
-            (this.options.$extended.offset().left > 348 ? 0 : 24)
-        , new_offset
-        ;
-
-      if (offset + height > window_height) {
-        if (height > window_height)
-          new_offset = 0;
-        else
-          new_offset = window_height - height;
-        this.options.$extended.css({ top: new_offset });
-      }
-    }
-
-  , get_tooltip_view: function(el) {
-      var tooltip_view = $(el)
-            .closest('[data-tooltip-view]').attr('data-tooltip-view');
-      if (tooltip_view && tooltip_view.length > 0) {
-        if (tooltip_view === "null") {
-          return null;
+      GGRC.Models.RecentlyViewedObject.findAll().done(function(objs) {
+        var max_history = that.options.max_history;
+        if(page_model) {
+          instance_list.push(new GGRC.Models.RecentlyViewedObject(page_model));
+          instance_list[0].save();
+          max_history--;
         }
-        else {
-          return GGRC.mustache_path + tooltip_view;
+
+        for(var i = objs.length - 1; i >= 0; i--) {
+          if((page_model && page_model.viewLink === objs[i].viewLink)
+            || objs.length - i > max_history || !("viewLink" in objs[i])
+          ) {
+            objs.splice(i, 1)[0].destroy(); //remove duplicate of current page object or excessive objects
+          } else if(instance_list.length < that.options.max_display) {
+            instance_list.push(objs[i]);
+          }
         }
-      }
-      else {
-        return this.options.tooltip_view;
-      }
-    }
 
-  , on_fade_in_timeout: function(el, instance) {
-      var self = this
-        , tooltip_view = this.get_tooltip_view(el)
-        ;
-
-      if (tooltip_view) {
-        this.fade_in_timeout = null;
-        can.view(tooltip_view, { instance: instance }, function(frag) {
-
-          var tooltip_width = self.options.$extended.outerWidth()
-            , offset = el.parent().offset()
-            , el_left = offset ? offset.left : 0
-            , offset_left = el_left - tooltip_width > 0 ?
-                el_left - tooltip_width : el_left + el.parent().width();
-
-          self.options.$extended
-            .html(frag)
-            .addClass('in')
-            .removeClass('hide')
-            .css({ top: el.offset().top, left: offset_left })
-            .data('model', instance);
-          self.ensure_tooltip_visibility();
+        can.view(that.options.list_view, {list : instance_list}, function(frag) {
+          that.element.find(".top-level.recent").html(frag);
         });
-      }
+      });
     }
-
-  , on_tooltip_mouseenter: function() {
-      clearTimeout(this.fade_out_timeout);
-      this.fade_out_timeout = null;
-    }
-
-  , on_fade_out_timeout: function() {
-      clearTimeout(this.fade_in_timeout);
-      this.fade_in_timeout = null;
-      this.fade_out_timeout = null;
-      this.options.$extended
-        .removeClass('in')
-        .addClass('hide')
-        .data('model', null);
-    }
-
-  , on_mouseleave: function(el, ev) {
-      // Cancel fade_in, if we haven't displayed yet
-      clearTimeout(this.fade_in_timeout);
-      this.fade_in_timeout = null;
-
-      clearTimeout(this.fade_out_timeout);
-      this.fade_out_timeout =
-        setTimeout(
-            this.proxy("on_fade_out_timeout"),
-            this.options.fade_out_delay);
-    }
-  , destroy: function() {
-    this._super();
-    this.on_mouseleave();
-  }
-});
-})(window.can, window.can.$);
+  });
